Trait reth_snapshot::segments::Segment

pub trait Segment: Default {
    // Required methods
    fn snapshot<DB: Database>(
        &self,
        provider: &DatabaseProviderRO<DB>,
        directory: impl AsRef<Path>,
        range: RangeInclusive<BlockNumber>
    ) -> ProviderResult<()>;
    fn segment(&self) -> SnapshotSegment;

    // Provided method
    fn dataset_for_compression<DB: Database, T: Table<Key = u64>>(
        &self,
        provider: &DatabaseProviderRO<DB>,
        range: &RangeInclusive<u64>,
        range_len: usize
    ) -> ProviderResult<Vec<Vec<u8>>> { ... }
}

A segment represents a snapshotting of some portion of the data.

Required Methods

fn snapshot<DB: Database>( &self, provider: &DatabaseProviderRO<DB>, directory: impl AsRef<Path>, range: RangeInclusive<BlockNumber> ) -> ProviderResult<()>

Snapshot data using the provided range. The directory parameter determines the snapshot file’s save location.

fn segment(&self) -> SnapshotSegment

Returns this struct’s SnapshotSegment.

Provided Methods

fn dataset_for_compression<DB: Database, T: Table<Key = u64>>( &self, provider: &DatabaseProviderRO<DB>, range: &RangeInclusive<u64>, range_len: usize ) -> ProviderResult<Vec<Vec<u8>>>

Generates the dataset to train a zstd dictionary with the most recent rows (at most 1000).

Object Safety

This trait is not object safe.